fix: ESLint エラー修正（no-explicit-any, react-hooks, react-refresh）
- SlashCommandList: useEffect内setState削除、commandの型をEditor/Rangeに変更
- getCommandItemsを別ファイル(slashCommandItems.tsx)に分離してreact-refresh対応
- MermaidBlock: transformPastedのany型をPmNode型に修正
