+++ src/components/tiptap/extensions/slashCommandItems.tsx
@@ -1,0 +1,94 @@
+import type { Editor, Range } from "@tiptap/core";
+import TitleIcon from "@mui/icons-material/Title";
+import FormatListBulletedIcon from "@mui/icons-material/FormatListBulleted";
+import FormatListNumberedIcon from "@mui/icons-material/FormatListNumbered";
+import CodeIcon from "@mui/icons-material/Code";
+import FormatQuoteIcon from "@mui/icons-material/FormatQuote";
+import HorizontalRuleIcon from "@mui/icons-material/HorizontalRule";
+import AccountTreeIcon from "@mui/icons-material/AccountTree";
+import type { CommandItem } from "./SlashCommandList";
+
+export function getCommandItems(): CommandItem[] {
+  return [
+    {
+      title: "Heading 1",
+      description: "Large section heading",
+      icon: <TitleIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).setHeading({ level: 1 }).run();
+      },
+    },
+    {
+      title: "Heading 2",
+      description: "Medium section heading",
+      icon: <TitleIcon fontSize="small" sx={{ fontSize: 18 }} />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).setHeading({ level: 2 }).run();
+      },
+    },
+    {
+      title: "Heading 3",
+      description: "Small section heading",
+      icon: <TitleIcon fontSize="small" sx={{ fontSize: 16 }} />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).setHeading({ level: 3 }).run();
+      },
+    },
+    {
+      title: "Bullet List",
+      description: "Unordered list",
+      icon: <FormatListBulletedIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).toggleBulletList().run();
+      },
+    },
+    {
+      title: "Numbered List",
+      description: "Ordered list",
+      icon: <FormatListNumberedIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).toggleOrderedList().run();
+      },
+    },
+    {
+      title: "Code Block",
+      description: "Code snippet",
+      icon: <CodeIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).toggleCodeBlock().run();
+      },
+    },
+    {
+      title: "Blockquote",
+      description: "Quote block",
+      icon: <FormatQuoteIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).toggleBlockquote().run();
+      },
+    },
+    {
+      title: "Horizontal Rule",
+      description: "Divider line",
+      icon: <HorizontalRuleIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor.chain().focus().deleteRange(range).setHorizontalRule().run();
+      },
+    },
+    {
+      title: "Mermaid Diagram",
+      description: "Flowchart, sequence diagram, etc.",
+      icon: <AccountTreeIcon fontSize="small" />,
+      command: ({ editor, range }: { editor: Editor; range: Range }) => {
+        editor
+          .chain()
+          .focus()
+          .deleteRange(range)
+          .insertContent({
+            type: "mermaidBlock",
+            attrs: { code: "graph TD\n    A[Start] --> B[End]" },
+          })
+          .run();
+      },
+    },
+  ];
+}
